Refactor signature configuration
Configure signature is now called as a part of the validate_signature?
method.  This allows models to customize the behavior if needed.

The valid_signature? method now receives authentication parameters to
use as it sees fit.

Split the configure_signature! method into pieces for easier
customization.  If we ever get around to writing more John Hancock
algorithms, this will need to be refactored again.